1. image-crawler수정. ( 이미지 년도, 이미지 사이즈, 기간 설정 )

+++ image-crawler.py
@@ -1,25 +1,39 @@
-from typing import Any, Union
-
-import cv2 as cv2
 import os
+import cv2 as cv2
+
 
 from icrawler.builtin import GoogleImageCrawler
 
-# google_crawler = GoogleImageCrawler(storage={'root_dir': 'animal/dog/'})
-# google_crawler.crawl('dog', max_num=15000, max_size=(1000,1000))
+
+folder_name = 'animal_train'
+start_year = 2010
+period = 1
+image_width = 224
+image_height = 224
+max_file_count = 1000
+
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
-dir_path += "/animal_train"
+dir_path += '/'
+dir_path += folder_name
 
 symbol = len(os.listdir(dir_path))
 
 classCount = 0
 for className in os.listdir(dir_path):
+    if className == '.DS_Store': continue
     classCount = 0
     fullPath = dir_path + "/" + className;
 
-    google_crawler = GoogleImageCrawler(storage={'root_dir': 'animal_train/' + className + '/'})
-    google_crawler.crawl(className, max_num=15000, min_size=(100, 100))
+    google_crawler = GoogleImageCrawler(storage={'root_dir': folder_name + '/' + className + '/'})
+
+    for year in range(period):
+        for month in range(11):
+            month += 1
+            google_crawler.crawl(keyword=className,
+                                 filters={'date': ((start_year + year,  month, 1), (start_year + year, month, 28))},
+                                 max_num=max_file_count,
+                                 file_idx_offset='auto')
 
     classCount = classCount + 1
     fileIndex = 0
@@ -31,22 +45,27 @@ for className in os.listdir(dir_path):
 
             filePath = fullPath + "/" + fileName
             filename, file_extension = os.path.splitext(filePath)
-            if file_extension != '.jpg':
+            if file_extension != '.jpg' and file_extension != '.bmp' and file_extension != '.gif' and file_extension != '.jpeg' and file_extension != '.png':
                 os.remove(filePath)
                 continue
 
             fileIndex = fileIndex + 1
 
-            modifiedPath = fullPath + "/" + str(fileIndex) + "_" + str(classCount) + "_" + className + file_extension
-            print(modifiedPath)
+            modifiedPath = fullPath + "/" + str(fileIndex) + "_" + str(classCount) + "_" + className + '.jpg'
+
             original = cv2.imread(filePath)
 
             if original is None:
                 os.remove(filePath)
-                fileIndex = fileIndex = - 1
+                fileIndex = fileIndex = fileIndex - 1
                 continue
 
-            resizeImage = cv2.resize(original, (224,224), interpolation=cv2.INTER_AREA)
+            if len(original.shape) != 3:
+                os.remove(filePath)
+                fileIndex = fileIndex = fileIndex - 1
+                continue
+
+            resizeImage = cv2.resize(original, (image_width,image_height), interpolation=cv2.INTER_AREA)
 
             cv2.imwrite(modifiedPath, resizeImage)
             cv2.imshow('resize', resizeImage)
